Working on setting data models
and access classes. Havent tested it the models now use model types instead of generic types that hold the foreign keys not sure if it will work but it makes more sense thisway. For example the first iteration stored a member variable called title that was a string this was to be used to link it to the affiliation type table but now I created a member variable that is of type affiliation and plan to access the content this way.