Got get all purchases to work
Needed to change the syntax of the query call also enabled some debugging tools with the show SQL in the hibernate.util file. Main issue was the orm class though wasnt matching the purchase table in the database so I had to line it up properly. Lastly I added more purchases so there would be more than one object in the list